@@ -2,8 +2,8 @@ package com.kaltura.playersdk;
 
 import android.content.Context;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.kaltura.playersdk.drm.DrmAdapter;
 import com.kaltura.playersdk.helpers.CacheManager;
