@@ -1,8 +1,8 @@
 package com.kaltura.playersdk.drm;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.kaltura.playersdk.LocalAssetsManager;
 
