@@ -4,8 +4,8 @@ import android.content.Context;
 import android.drm.DrmErrorEvent;
 import android.drm.DrmEvent;
 import android.drm.DrmInfoEvent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.kaltura.playersdk.LocalAssetsManager;
