@@ -1,7 +1,7 @@
 package com.kaltura.playersdk.helpers;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
